backend/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
pyjwt[cryptography]
python-multipart
email-validator
